requirements.txt
ccxt
ipykernel
jupyter
lightgbm
matplotlib
numpy
pandas
pyarrow
python-docx
scikit-learn
TA-Lib
torch
xgboost